test: extract RecordingTestSupport (MockAudioRecorder) from TestSupport (#550, #434)
First slice of splitting the 1,048-line TestSupport.swift test "junk
drawer" into domain files, one cohesive mock group per PR, names stable,
green between (#434 plan; unblocked now that #431 is merged).

Moves the recording mock — the @MainActor MockAudioRecorder
(AudioRecording, MicrophonePermissionAccessing) plus its nested
ActivationProbeBehavior — byte-identical into a new
RecordingTestSupport.swift with the standard test imports. TestSupport.swift
keeps every other mock unchanged.

Pure test-support move: no product code, no helper renames, no behavior
change. The full BugNarratorTests suite stays green at 662 tests
(unchanged count — nothing added or removed); docs-sync validation
passes.

Pre-build spec reviewed by codex in a fresh session (confirmed the mock
is self-contained; 2 findings folded: Xcode target membership via
xcodegen, and a concrete 662-test baseline for the unchanged-count gate).
